Struct libcad::llir::interp::Interp

pub struct Interp { /* fields omitted */ }

An interpreter.

Implementations

impl Interp

pub fn new(datalayout: DataLayout, typedefs: Typedefs<ExtIdent>) -> Interp

Returns a new interpreter with the given data layout and typedef list.

pub fn datalayout(&self) -> &DataLayout

Returns the data layout.

pub fn typedefs(&self) -> &Typedefs<ExtIdent>

Returns the typedef list.

pub fn func(&self, name: &GlobalIdent) -> Result<&Func, Error>

Returns the function by name.

pub fn define_func(
&mut self,
name: GlobalIdent,
func: Func
) -> Result<(), Error>

Defines the function.

pub fn global(&self, name: &GlobalIdent) -> Result<Val, Error>

Returns the value of the global variable.

pub fn define_global(
&mut self,
name: GlobalIdent,
obj: Val
) -> Result<(), Error>

Defines the global variable with the value.

pub fn alloca(&mut self, val: Val) -> Region

Returns a new alloca region with the value.

pub fn deref_alloca(&self, id: usize, offset: i64) -> Result<ValView, Error>

Returns the value view by alloca-ed index and offset.

pub fn store_to_alloca(&mut self, v: Val, id: usize) -> Result<(), Error>

Stores the value view to the alloca-ed region with the given index.

pub fn malloc(&mut self, size: i64) -> Result<Region, Error>

Returns a new malloc region.

pub fn deref_heap(&self, id: usize, offset: i64) -> Result<ValView, Error>

Returns the value view by malloc-ed index and offset.

pub fn store_to_heap(&mut self, v: Val, id: usize) -> Result<(), Error>

Stores the value view to the malloc-ed region with the given index.

pub fn deref_addr(&self, addr: &Addr) -> Result<ValView, Error>

Returns the value view by dereferences the given address.

pub fn bitcast_poison(
&mut self,
addr: &Addr,
ty: &LLIRType<ExtIdent>
) -> Result<(), Error>

Collapses the poison region at the given address with the given type.

pub fn store_to(&mut self, val: Val, region: &Region) -> Result<(), Error>

Stores the value to the region.

pub fn local(&self, name: &LocalIdent) -> Result<&Val, Error>

Returns the value of the local variable.

pub fn set_local(&mut self, name: LocalIdent, obj: Val) -> Result<(), Error>

Sets the value to the local variable.

pub fn push_binop_value(
&self,
view: &mut ValView,
opcode: BinOpcode,
view1: &ValView,
view2: &ValView,
pad: bool
) -> Result<(), Error>

Pushes the result of binary operation.

pub fn push_cmpop_value(
&self,
view: &mut ValView,
opcode: CmpOpcode,
view1: &ValView,
view2: &ValView,
pad: bool
) -> Result<(), Error>

Pushes the result of comparison operation.

pub fn push_value(
&self,
view: &mut ValView,
ty: Option<&LLIRType<ExtIdent>>,
val: &Value<ExtIdent>,
pad: bool
) -> Result<(), Error>

Pushes the value.

pub fn eval_value(
&self,
ty: Option<&LLIRType<ExtIdent>>,
val: &Value<ExtIdent>
) -> Result<Val, Error>

Returns an evaluated value.

pub fn eval_insn(&self, insn: &Insn<ExtIdent>) -> Result<Effect, Error>

Evaluates the instruction and returns the caused effect.

pub fn stacktrace(&self) -> Result<Vec<Loc>, Error>

Returns the stack trace.

pub fn state(&self) -> &State

Returns the state.

pub fn insn(&self) -> Result<&Insn<ExtIdent>, Error>

Returns the next instruction.

pub fn enter(
&mut self,
res: Option<LocalIdent>,
funcname: GlobalIdent,
args: Vec<Val>
) -> Result<(), Error>

Enters into a new call stack frame by the function call.

pub fn leave(&mut self, ret: Option<Val>) -> Result<(), Error>

Leaves from the current call frame.

pub fn start(&mut self, funcname: GlobalIdent) -> Result<(), Error>

Starts the execution by calling the function without any arguments.

pub fn step(&mut self) -> Result<Effect, Error>

Steps the interpreter.