Module libcad::llir::interp::syntax

The syntax of the LLIR interpreter.

Structs

Addr

An address that is a pair of region and offset.

Func

A function that is the tuple of name, arguments and blocks.

Val

A value that is the sequence of constants.

Enums

Const

A constant.

Cont

A continuation representing control flow update.

Effect

An effect representing control flow or heap update.

Error

An error generated by an interpreter.

PoisonVal

A value that may be poisoned.

Region

A region of Addr.

State

A state of an interpreter.