Module libcad::llir::syntax

The syntax of LLIR.

Structs

BinOpArgs

Arguments of binary operation (https://llvm.org/docs/LangRef.html#binary-operations).

Block

A block of Blocks.

Blocks

A list of blocks of a function body (https://llvm.org/docs/LangRef.html#functions).

CallBody

An argument body of call instruction (https://llvm.org/docs/LangRef.html#call-instruction).

CmpOpArgs

Arguments of a comparison operation.

ConvOpArgs

Arguments of a conversion operation (https://llvm.org/docs/LangRef.html#conversion-operations).

DIFile

A DIFile (https://llvm.org/docs/LangRef.html#difile).

DILexicalBlock

A DILexicalBlock (https://llvm.org/docs/LangRef.html#dilexicalblock).

DILexicalBlockFile

A DILexicalBlockFile (https://llvm.org/docs/LangRef.html#dilexicalblockfile).

DILocation

A DILocation (https://llvm.org/docs/LangRef.html#dilocation).

DISubprogram

A DISubprogram (https://llvm.org/docs/LangRef.html#disubprogram).

ExtIdentUnit

The unit extension identifier.

FuncDecl

A declaration of a function (https://llvm.org/docs/LangRef.html#functions).

FuncSig

A signature of a function (https://llvm.org/docs/LangRef.html#functions).

GetelementptrArgs

Arguments of getelementptr operation (https://llvm.org/docs/LangRef.html#getelementptr-instruction).

GlobalIdent

A global identifier (https://llvm.org/docs/LangRef.html#identifiers).

Label

A label.

LineInfo

A line number information.

Loc

A location of an instruction.

LocLineInfo

A location of an instruction with a line number information.

LocalIdent

A local identifier (https://llvm.org/docs/LangRef.html#identifiers).

MetadataList

A metadata list that is a map from the string to the metadata reference.

MetadataRef

A metadata reference (https://llvm.org/docs/LangRef.html#metadata).

Module

A module of LLIR.

Param

A parameter (https://llvm.org/docs/LangRef.html#functions).

ParamAttr

An entry of ParamAttrs.

ParamAttrs

Parameter attributes (https://llvm.org/docs/LangRef.html#parameter-attributes).

ParamValue

A parameter value.

RetParam

A return parameter (https://llvm.org/docs/LangRef.html#functions).

TypedValue

A typed value.

Typedefs

A typedef list that is a map from the local identifier of the variable to the type.

UnOpArgs

Arguments of a unary operation (https://llvm.org/docs/LangRef.html#unary-operations).

Enums

AggOpArgs

An aggregate operation (https://llvm.org/docs/LangRef.html#aggregate-operations).

Attr

An entry of an attribute group (https://llvm.org/docs/LangRef.html#attribute-groups).

BinOpcode

An opcode of BinOpArgs.

CallConv

A calling convention of CallBody.

Callee

A callee value of call instruction (https://llvm.org/docs/LangRef.html#call-instruction).

CmpOpcode

An opcode of CmpOpArgs.

ConvOpcode

A opcode of ConvOpArgs.

FcmpCond

A condition of fcmp (https://llvm.org/docs/LangRef.html#fcmp-instruction).

IcmpCond

A condition of icmp (https://llvm.org/docs/LangRef.html#icmp-instruction).

Insn

An instruction (https://llvm.org/docs/LangRef.html#instruction-reference).

LandingpadClause

A clause of landingpad instruction (https://llvm.org/docs/LangRef.html#landingpad-instruction).

Linkage

A linkage type (https://llvm.org/docs/LangRef.html#linkage-types).

Metadata

A metadata (https://llvm.org/docs/LangRef.html#metadata).

ParamAttrArg

An argument of ParamAttr.

Stmt

A statement of LLIR.

TailMode

A tail call mode of call instruction (https://llvm.org/docs/LangRef.html#call-instruction).

Type

A type (https://llvm.org/docs/LangRef.html#type-system).

UnOpcode

An opcode of UnOpArgs.

UnnamedAddr

An unnamed_addr specifier (https://llvm.org/docs/LangRef.html#global-variables) and FuncDecl.

Value

A value (https://llvm.org/docs/LangRef.html#constants, https://llvm.org/docs/LangRef.html#other-values). This also contains the composable instructions without side effects.

VecOpArgs

Arguments of a vector operation (https://llvm.org/docs/LangRef.html#vector-operations).

Visibility

A visibility style (https://llvm.org/docs/LangRef.html#visibility-styles).

WrapMode

A wrap mode of BinOpcode.

Traits

ExtIdent

A trait for an extension identifier.