Module libcad::typechk::cast

The casting analysis module.

Structs

CastChk

A cast checker.

CastChkResult

A result of cast checking.

CastChkState

A state of CastChk.

DerefValue

A (dereferenced) value.

FieldVar

A field variable.

FieldVarBase

A base pointer of FieldVar.

Path

A path that is the list of the locations.

Resolver

A resolver of multiple results.

StoreResult

A result of CondState::Store.

Enums

CastChkKind

A kind of the cast checking.

CastChkResultKind

A kind of CastChkKind.

CondState

A state of extracting a condition.

PathState

A state of Path.

TargetState

A state of identifying a target pointer.

VarState

A state of identifying the pointer of a variable.