Enum libcad::typechk::cast::CondState

pub enum CondState {
    SearchBrTo(Option<Label>),
    SearchStoreTo(ExtIdent),
    CondVar {
        negated: bool,
        var: LocalIdent,
    },
    Store {
        loc: Loc,
        ptr: PtrExtIdent,
        ty: LLIRType<ExtIdent>,
        src: Box<CondState>,
        dst: Box<CondState>,
    },
    BinOp {
        opcode: BinOpcode,
        left: Box<CondState>,
        right: Box<CondState>,
    },
    CmpOp {
        opcode: CmpOpcode,
        left: Box<CondState>,
        right: Box<CondState>,
    },
    ConstInt(usize, i64),
    Deref(Box<CondState>),
    Var(TargetState, LocalIdent),
    VarState(VarState),
    UnknownValue(Value<ExtIdent>, Loc),
    Failed(Box<CondState>, Loc),
}

A state of extracting a condition.

Variants

SearchBrTo(Option<Label>)

Searching an branch instruction to the block.

Tuple Fields of SearchBrTo

0: Option<Label>

SearchStoreTo(ExtIdent)

Searching an store instruction to a value of the refinement type.

Tuple Fields of SearchStoreTo

0: ExtIdent

CondVar

Extracting a (optionally negated) condition assigned to the variable.

Fields of CondVar

negated: boolvar: LocalIdent

Store

Storing a value with the type to the pointer with the refinement type at the location.

Fields of Store

loc: Locptr: PtrExtIdentty: LLIRType<ExtIdent>src: Box<CondState>dst: Box<CondState>

BinOp

Applying the binary operation.

Fields of BinOp

opcode: BinOpcodeleft: Box<CondState>right: Box<CondState>

CmpOp

Applying the comparison operation.

Fields of CmpOp

opcode: CmpOpcodeleft: Box<CondState>right: Box<CondState>

ConstInt(usize, i64)

A constant n-bit integer (n <= 64).

Tuple Fields of ConstInt

0: usize1: i64

Deref(Box<CondState>)

Dereferencing a value of the condition.

Tuple Fields of Deref

0: Box<CondState>

Var(TargetState, LocalIdent)

Tracing the target variable.

Tuple Fields of Var

0: TargetState1: LocalIdent

VarState(VarState)

A state of tracing the variable.

Tuple Fields of VarState

0: VarState

UnknownValue(Value<ExtIdent>, Loc)

Detected unknown value at the location.

Tuple Fields of UnknownValue

0: Value<ExtIdent>1: Loc

Failed(Box<CondState>, Loc)

Failed on the state at the location.

Tuple Fields of Failed

0: Box<CondState>1: Loc

Implementations

impl CondState

pub fn from_value(
target: TargetState,
base: DerefValue<'_>,
loc: &Loc
) -> CondState

Returns a new state for a variable or a value.

pub fn varstate(&self) -> Option<&VarState>

Returns the contained VarState.

pub fn is_invalid_store(&self, env: &TypeEnv) -> bool

Returns true if the state is valid store.

pub fn try_to_store_on(
&self,
base: LocalIdent,
rebaseoff: i64
) -> Option<StoreResult<'_>>

Returns the result of the store on the given base variable with the rebase offset.

pub fn update(
&mut self,
insn: &Insn<ExtIdent>,
effect: &Effect,
loc: &Loc,
env: &TypeEnv
) -> Result<bool, String>

Updates the state with the instruction at the location under the effect and the typing environment.

pub fn try_to_cond_const(&self) -> Option<CondConst>

Returns an extracted constant in the expression in the condition.

pub fn try_to_cond_expr(
&self,
varmap: &mut VarMap<FieldVar>
) -> Option<CondExpr>

Returns an extracted expression in the condition.

pub fn to_cond(&self, varmap: &mut VarMap<FieldVar>) -> Cond

Returns an extracted condition.