Struct libcad::typechk::cast::CastChk

pub struct CastChk { /* fields omitted */ }

A cast checker.

Implementations

impl CastChk

pub fn from_cast_reason(reason: &CastReason, loc: &Loc) -> CastChk

Returns a new cast checker from the cast reason and the location.

pub fn check_annot_file_validity(
annotfile: &AnnotFile,
env: &TypeEnv
) -> Result<(), String>

Check the validity of the annotation file under the typing environment.

pub fn judge_cast(
ty1: &LLIRType<ExtIdent>,
ty2: &LLIRType<ExtIdent>,
loc: &Loc,
env: &TypeEnv
) -> Result<bool, String>

Judges the type cast at the location under the type environment by TypeEnv::judge_cast.

pub fn judge_cast_match_target(
cases: &[CastMatchCase],
reason: CastReasonKind,
cond: &Cond,
target: &LLIRType<ExtIdent>,
loc: &Loc,
env: &TypeEnv
) -> Option<Cond>

Judges the type cast at the location under the given cast-match and the condition under the type environment.

pub fn judge_cast_match_equality(
refiner: &ApplyRefiner,
cases: &[CastMatchCase],
cond: &Cond
) -> Option<Cond>

Judges the equality of the condition under the given cast-match and apply-refiner.

pub fn state(&self) -> &CastChkState

Returns the state.

pub fn ctx(&self) -> &Cond

Returns the context condition.

pub fn matched_cond(&self) -> &Option<Cond>

Returns the matched condition.

pub fn srcs(&self) -> &[(Loc, PtrExtIdent)]

Returns the list of the source location and pointer extension identifier.

pub fn chk_kind(&self) -> &CastChkKind

Returns the kind.

pub fn run(
self,
annotfile: &AnnotFile,
env: &TypeEnv,
module: &Module<ExtIdent>
) -> CastChkResult

Runs the automaton under the annotation file, typing environment and LLIR module.