Struct libcad::typechk::env::TypeEnv

pub struct TypeEnv { /* fields omitted */ }

A typing environment which implements the structural typing.

Implementations

impl TypeEnv

pub fn empty(datalayout: DataLayout) -> TypeEnv

Returns a new typing environment.

pub fn datalayout(&self) -> &DataLayout

Returns the datalayout.

pub fn typedefs(&self) -> &Typedefs<ExtIdent>

Returns the typedef list.

pub fn typeset(&self) -> &BTreeMap<LocalIdent, Type>

Returns the typeset (interned typedef list).

pub fn varset(&self) -> &BTreeMap<VarName, Type>

Returns the varset (the map from the variable name to the bound type).

pub fn ty(&self, name: &LocalIdent) -> Result<&Type, String>

Returns the type in the typeset by name.

pub fn define_type(
&mut self,
name: LocalIdent,
prototype: LLIRType<ExtIdent>,
ty: Type
) -> Result<(), String>

Defines the type with the name and the LLIR prototype.

pub fn func(&self, name: &GlobalIdent) -> Result<&FuncArgs, String>

Returns the function by name.

pub fn declare_func(
&mut self,
name: GlobalIdent,
ty: Type,
args: FuncArgs
) -> Result<(), String>

Declares the function with the name, type and arguments.

pub fn var(&self, name: &VarName) -> Result<&Type, String>

Returns the type bound to the variable by name.

pub fn declare_var(&mut self, name: VarName, ty: Type) -> Result<(), String>

Declares the variable with the name and type.

pub fn effectset(&self) -> &BTreeMap<Loc, Effect>

Returns the effect set (the map from the location to the effect).

pub fn effect(&self, loc: &Loc) -> Option<&Effect>

Returns the effect by the location.

pub fn insert_effect(&mut self, loc: Loc, effkind: EffectKind)

Inserts the effect at the location.

pub fn new_infervar(&mut self) -> InferVar

Returns a new (fresh) inference variable.

pub fn alignof_type(&self, ty: &LLIRType<ExtIdent>) -> Option<usize>

Returns the alignment of the LLIR type via DataLayout::alignof_type.

pub fn sizeof_type(&self, ty: &LLIRType<ExtIdent>) -> Option<usize>

Returns the size of the LLIR type via DataLayout::sizeof_type.

pub fn intern_llir_type(
&self,
ty: &LLIRType<ExtIdent>,
loc: &Loc
) -> Result<Type, String>

Returns an type interned from the given LLIR type at the location.

pub fn intern_refresh_llir_type(
&mut self,
ty: &LLIRType<ExtIdent>,
all: bool,
loc: &Loc
) -> Result<Type, String>

Returns a type interned from the given LLIR type at the location. The returned type is refreshed via Type::refresh.

pub fn intern_fresh_llir_type(
&mut self,
ty: &LLIRType<ExtIdent>,
loc: &Loc
) -> Result<Type, String>

Calls TypeEnv::intern_refresh_llir_type with all = false.

pub fn intern_type_value(
&mut self,
mode: InternMode,
ty: Option<&LLIRType<ExtIdent>>,
val: &Value<ExtIdent>,
loc: &Loc,
constrs: &mut Constrs
) -> Result<Type, String>

Returns a type interned from the given LLIR type and value at the location.

pub fn intern_typedvalue(
&mut self,
mode: InternMode,
tyval: &TypedValue<ExtIdent>,
loc: &Loc,
constrs: &mut Constrs
) -> Result<Type, String>

Returns a type interned from the given LLIR typed value at the location.

pub fn intern_value_expr(
&mut self,
mode: InternMode,
val: &Value<ExtIdent>,
loc: &Loc,
constrs: &mut Constrs
) -> Result<Type, String>

Returns a type interned from the given value at the location.

pub fn judge_alloc(
&mut self,
kind: &AllocKind,
ty: &Type,
loc: &Loc,
constrs: &mut Constrs
) -> Result<(), String>

Judges the allocation.

pub fn intern_value_ext(
&mut self,
mode: InternMode,
val: ValueExt,
loc: &Loc,
constrs: &mut Constrs
) -> Result<Type, String>

Returns a type interned from the extended value at the location.

pub fn judge_subtype(
&self,
ty1: &Type,
ty2: &Type,
base_differs: bool,
loc: &Loc,
constrs: &mut Option<&mut Constrs>,
stack: &mut Vec<(Type, Type)>
) -> Result<bool, String>

Judges whether ty1 is subtype of ty2 at the location. See TypeEnv::judge_cast.

pub fn judge_cast(
&self,
ty1: &Type,
ty2: &Type,
base_differs: bool,
loc: &Loc,
constrs: &mut Option<&mut Constrs>
) -> Result<bool, String>

Judges whether ty1 can be cast as ty2 at the location. If base_differs is true, base types do not have the subtype relation each other. stack is internally used for avoiding infinite loop.

pub fn judge_term(
&mut self,
term: JudgeTerm,
loc: &Loc,
constrs: &mut Constrs
) -> Result<(), String>

Judges the given term at the location.

pub fn judge_overwrite(
&self,
target: &Type,
ty: &Type,
loc: &Loc
) -> Result<bool, String>

Judges whether values with the target type may be overwritten by values with the type at the location.

pub fn judge_effect_modifies(
&self,
effect: &Effect,
target: &Type,
loc: &Loc
) -> bool

Judges whether values with the target types may be modified by the effect at the location.

pub fn judge_effect_modifies_ptr(
&self,
effect: &Effect,
target: &Type,
loc: &Loc
) -> bool

Judges whether pointers with the target type may be modified by the effect at the location.