Module libcad::typechk::syntax

The syntax of the types.

Structs

Constrs

Constraints that is the list of the pair of the instruction location and the constraint.

Effect

An effect that is a list of the pair of the effect kind and the instruction location.

FuncArgs

Arguments of a function.

InferVar

An inference variable used in identifying pointers.

PtrExtIdent

A pointer extention identifier that is an extention identifier with a inference variable of a pointer.

PtrExtIdentDummy

A pointer extention identifier with a dummy inference variable of a pointer.

StoreEffect

A store (memory copy) effect.

Enums

AllocKind

A kind of memory allocation.

AllocReason

A reason of Constr::Alloc.

CastReason

A cast reason indicating the reason for checking the cast.

CastReasonKind

A kind of CastReason.

Constr

A constraint which will be resolved after the structural type checking.

ConstrKind

A kind of Constr.

EffectKind

A kind of Effect.

FreeKind

A kind of Constr::Free.

JudgeTerm

A term in judgement.

Size

A size in byte.

Type

A type.

ValueExt

An extended value.

VarName

A name of a variable.